Update zip documentation
The example of sorting a zipped range was very confusing as it used a string literal sorting predicate, requiring use of "a" and "b" which were already being used as the names of the two ranges being zipped.  Updated the sorting predicate to use a lambda with parameters "c" and "d".